Stretch Activity from team wk3 assignment: Refactored the shopping cart
